Module Caqti_lwt_unix.Pool

type ('a, +'e) t

val size : ('a, 'e) t -> int

size pool is the current number of open resources in pool.

val use : ?priority:float -> ('a -> ('b, 'e) Stdlib.result Lwt.t) -> ('a, 'e) t -> ('b, 'e) Stdlib.result Lwt.t

use f pool calls f on a resource drawn from pool, handing back the resource to the pool when f exits.

parameter priority
Requests for the resource are handled in decreasing order of priority. The default priority is 0.0.

val drain : ('a, 'e) t -> unit Lwt.t

drain pool closes all resources in pool. The pool is still usable, as new resources will be created on demand.