Module Caqti_lwt

Prerequisities for connecting to databases using Lwt

This module contains most of the prerequisite types and modules. Functions to establish database connections are provided by the caqti-lwt.unix and caqti-mirage libraries. Pool instances are also found there due to additional OS dependencies.

module Stream : Caqti_stream_sig.S with type 'a fiber := 'a Lwt.t

module Switch : Caqti_switch_sig.S with type 'a fiber := 'a Lwt.t

module type CONNECTION = Caqti_connection_sig.S with type 'a fiber := 'a Lwt.t and type ('a, 'e) stream := ('a, 'e) Stream.t

type connection = (module CONNECTION)

val or_fail : ('a, [< Caqti_error.t ]) Stdlib.result -> 'a Lwt.t

Converts an error to an Lwt future failed with a Caqti_error.Exn exception holding the error.